chore: improve the Url component
