+++ blog/2024-01-02-vel-release/index.mdx
@@ -10,7 +10,7 @@ import JoinComm from "@site/src/components/JoinComm";
 import GHUrl from "@site/src/components/GHUrl";
 import ReleaseNotes from "@site/src/components/ReleaseNotes";
 import GitHubProject from "@site/src/components/GitHubProject";
-import PR from "@site/src/components/PR";
+import IssueOrPr from "@site/src/components/IssueOrPr";
 import Details from "@site/src/components/Details";
 
 ## <span className="color-section">New Velaptor Release</span>
@@ -270,7 +270,7 @@ This means the dictionary doesn't need to be resized as more keys are added when
 These strategies, among others, not only improved performance but also significantly reduced memory allocations.
 
 :::note Code changes
-Curious about the nitty-gritty details? Feel free to explore the <PR projName="Velaptor" prNum="849" urlText="pull request here"/>
+Curious about the nitty-gritty details? Feel free to explore the <IssueOrPr repoName="Velaptor" isIssue={false} issueOrPrNumber="849" urlText="pull request here"/>
 to see exactly how we improved the performance of the keyboard input system.
 :::
